+++ earthfile2llb/interpreter_test.go
@@ -28,48 +28,3 @@ func TestBuildArgMatrix(t *testing.T) {
 		assert.Equal(t, tt.out, ans)
 	}
 }
-
-func TestParseParams(t *testing.T) {
-	var tests = []struct {
-		in    string
-		first string
-		args  []string
-	}{
-		{"(+target/art --flag=something)", "+target/art", []string{"--flag=something"}},
-		{"(+target/art --flag=something\"\")", "+target/art", []string{"--flag=something\"\""}},
-		{"( \n  +target/art \t \n --flag=something\t   )", "+target/art", []string{"--flag=something"}},
-		{"(+target/art --flag=something\\ --another=something)", "+target/art", []string{"--flag=something\\ --another=something"}},
-		{"(+target/art --flag=something --another=something)", "+target/art", []string{"--flag=something", "--another=something"}},
-		{"(+target/art --flag=\"something in quotes\")", "+target/art", []string{"--flag=\"something in quotes\""}},
-		{"(+target/art --flag=\\\"something --not=in-quotes\\\")", "+target/art", []string{"--flag=\\\"something", "--not=in-quotes\\\""}},
-		{"(+target/art --flag=look-ma-a-\\))", "+target/art", []string{"--flag=look-ma-a-\\)"}},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.in, func(t *testing.T) {
-			actualFirst, actualArgs, err := parseParams(tt.in)
-			assert.NoError(t, err)
-			assert.Equal(t, tt.first, actualFirst)
-			assert.Equal(t, tt.args, actualArgs)
-		})
-
-	}
-}
-
-func TestNegativeParseParams(t *testing.T) {
-	var tests = []struct {
-		in string
-	}{
-		{"+target/art --flag=something)"},
-		{"(+target/art --flag=something"},
-		{"(+target/art --flag=\"something)"},
-		{"(+target/art --flag=something\\)"},
-		{"()"},
-		{"(          \t\n   )"},
-	}
-
-	for _, tt := range tests {
-		_, _, err := parseParams(tt.in)
-		assert.Error(t, err)
-	}
-}
